fix fmt string issues

+++ src/exchange-lib/exchange_api_deposit_wtid.c
@@ -224,7 +224,7 @@ handle_deposit_wtid_finished (void *cls,
     /* unexpected response code */
     GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                 "Unexpected response code %u\n",
-                response_code);
+                (unsigned int) response_code);
     GNUNET_break (0);
     response_code = 0;
     break;
