+++ src/exchange-lib/exchange_api_handle.c
@@ -632,7 +632,7 @@ keys_completed_cb (void *cls,
   default:
     GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                 "Unexpected response code %u\n",
-                response_code);
+                (unsigned int) response_code);
     break;
   }
