+++ src/exchange-lib/exchange_api_refresh_link.c
@@ -364,7 +364,7 @@ handle_refresh_link_finished (void *cls,
     /* unexpected response code */
     GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                 "Unexpected response code %u\n",
-                response_code);
+                (unsigned int) response_code);
     GNUNET_break (0);
     response_code = 0;
     break;
